calibrated IR sensor

+++ HighLevelDrivers/inc/IRSensor.h
@@ -1,4 +1,5 @@
 #include "../../LowLevelDrivers/inc/adc.h"
 
-
-uint16_t get_IR_distance(ADC12_Regs *adc12);
+// Returns the measured distance in mm
+// Note that measurements less than 80mm are unreliable, and the maximum distance is about 150mm
+uint8_t get_IR_distance(ADC12_Regs *adc12);
